+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [3.2.0] - 2020-01-22
 ### Fixed
 - Allow to store bytes content using Mock.
 - Mock is now able to retrieve a file that was previously stored.
@@ -16,5 +18,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/Colin-b/pyndows/compare/v3.1.0...HEAD
+[Unreleased]: https://github.com/Colin-b/pyndows/compare/v3.2.0...HEAD
+[3.2.0]: https://github.com/Colin-b/pyndows/compare/v3.1.0...v3.2.0
 [3.1.0]: https://github.com/Colin-b/pyndows/releases/tag/v3.1.0